# src/main/resources/api_docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Maize disease detection API Reference",
    "description": "This api  houses all the logic that is required when interacting the maize leaf disease detection mobile clientIt contains mainly four parts. \n1. User account\n2. Diseases api\n3. Neural network model\n4. Community section",
    "contact": {
      "name": "Augustine Simwela",
      "url": "https://github.com/ASIMWELA",
      "email": "augastinesimwela@gmail.com"
    },
    "version": "API Version 1.0"
  },
...
  "paths": {
...
    "/api/v1/diseases": {
      "get": {
        "tags": [
          "Diseases Controller"
        ],
        "summary": "Get paginated diseases",
        "operationId": "getDiseases",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "pagedResourceAssembler",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PagedResourcesAssemblerDiseaseEntity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelObject"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Diseases Controller"
        ],
        "summary": "Adds a disease to the database",
        "operationId": "saveDisease",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiseaseEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/diseases/symptoms/{diseaseUuid}": {
      "post": {
        "tags": [
          "Diseases Controller"
        ],
        "summary": "Add symptoms of a particular disease",
        "operationId": "addDiseaseSymptoms",
        "parameters": [
          {
            "name": "diseaseUuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddSymptomRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/diseases/prescriptions/{diseaseUuid}": {
      "post": {
        "tags": [
          "Diseases Controller"
        ],
        "summary": "Add prescriptions of a particular disease",
        "operationId": "addDiseasePrescription",
        "parameters": [
          {
            "name": "diseaseUuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddPrescriptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/diseases/{diseaseUuid}": {
      "get": {
        "tags": [
          "Diseases Controller"
        ],
        "summary": "Get a single disease with its symptoms and prescriptions",
        "operationId": "getDisease",
        "parameters": [
          {
            "name": "diseaseUuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GetDiseaseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/diseases/get-symptoms/{diseaseUuid}": {
      "get": {
        "tags": [
          "Diseases Controller"
        ],
        "summary": "Get disease symptoms.Given the disease uuid, return the symptoms if the disease is found in the database otherwise throw 404 exception",
        "operationId": "getDiseaseSymptoms",
        "parameters": [
          {
            "name": "diseaseUuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionModelObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/diseases/get-prescriptions/{diseaseUuid}": {
      "get": {
        "tags": [
          "Diseases Controller"
        ],
        "summary": "Get a disease prescriptions. Given a disease uuid, return the prescriptions of that particular disease if the disease is found otherwise throw 404",
        "operationId": "getDiseasePrescriptions",
        "parameters": [
          {
            "name": "diseaseUuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionModelObject"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Links": {
        "type": "object",
        "additionalProperties": {
          "$ref": "#/components/schemas/Link"
        }
      },
...
      "PageMetadata": {
        "type": "object",
        "properties": {
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int64"
          },
          "number": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "PagedModelObject": {
        "type": "object",
        "properties": {
          "_embedded": {
            "type": "object",
            "properties": {
              "objects": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              }
            }
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          },
          "page": {
            "$ref": "#/components/schemas/PageMetadata"
          }
        }
      },
...
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "GetDiseaseResponse": {
        "type": "object",
        "properties": {
          "diseaseUuid": {
            "type": "string"
          },
          "diseaseName": {
            "type": "string"
          },
          "accuracy": {
            "type": "string"
          },
          "symptoms": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "prescriptions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
      "DiseaseEntity": {
        "required": [
          "diseaseName"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "uuid": {
            "type": "string"
          },
          "diseaseName": {
            "type": "string"
          }
        }
      },
      "AddSymptomRequest": {
        "required": [
          "symptoms"
        ],
        "type": "object",
        "properties": {
          "symptoms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SymptomEntity"
            }
          }
        }
      },
      "SymptomEntity": {
        "required": [
          "symptomDescription"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "uuid": {
            "type": "string"
          },
          "symptomDescription": {
            "type": "string"
          }
        }
      },
      "AddPrescriptionRequest": {
        "required": [
          "prescriptions"
        ],
        "type": "object",
        "properties": {
          "prescriptions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PrescriptionEntity"
            }
          }
        }
      },
      "PrescriptionEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "uuid": {
            "type": "string"
          },
          "diseasePrescription": {
            "type": "string"
          }
        }
      },
...
      "PagedResourcesAssemblerDiseaseEntity": {
        "type": "object",
        "properties": {
          "forceFirstAndLastRels": {
            "type": "boolean",
            "writeOnly": true
          }
        }
      },
      "CollectionModelObject": {
        "type": "object",
        "properties": {
          "_embedded": {
            "type": "object",
            "properties": {
              "objects": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              }
            }
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
...
      "Link": {
        "type": "object",
        "properties": {
          "href": {
            "type": "string"
          },
          "hreflang": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "deprecation": {
            "type": "string"
          },
          "profile": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "templated": {
            "type": "boolean"
          }
        }
      }
    },
    "securitySchemes": {
      "api": {
        "type": "http",
        "in": "header",
        "scheme": "basic"
      }
    }
  }
}